test: first federation integration test
The simplest test I could imagine: querying a single field from a single
subgraph. This fails at the moment so I've `#[ignore]`d it, PRing so it can be
used to aid in the implementation.

Part of GB-5337
